Add classes for constructing parallel text searcher from gzip with recovery points

+++ ArchiveViewerBackend/WithByteOffset.cs
@@ -1,10 +1,10 @@
 ﻿namespace ArchiveViewerBackend;
 
-public record WithByteOffset<T>(T Payload, int ByteOffset);
+public record WithByteOffset<T>(T Payload, long ByteOffset);
 
 public static class WithByteOffsetExtensions
 {
-    public static WithByteOffset<T> WithByteOffset<T>(this T payload, int byteOffset)
+    public static WithByteOffset<T> WithByteOffset<T>(this T payload, long byteOffset)
     {
         return new WithByteOffset<T>(payload, byteOffset);
     }
